api/requirements.txt
flask>=3.0
flask-cors>=4.0
gunicorn>=21.0
pillow>=10.0
